requirements.txt
# FastAPI y ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# GraphQL
strawberry-graphql[fastapi]==0.219.0

# Base de datos
sqlalchemy==2.0.25
asyncpg==0.29.0  # Para PostgreSQL async
alembic==1.13.1
greenlet==3.0.3. # Simular el puente entre sync/async
psycopg2-binary>=2.9  # Conectarse modo sync a PostgreSQL (para Alembic)

# Autenticación
pyjwt==2.8.0
bcrypt==4.1.2
passlib==1.7.4

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# Configuración
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilidades
python-dateutil==2.8.2

# Testing (opcional)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0